requirements.txt
# Core
python-dotenv>=1.0.0
ollama>=0.4.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.34.0

# Email
imapclient>=3.0.0
exchangelib>=5.4.0

# PO Parsing
pdfplumber>=0.11.0
openpyxl>=3.1.0
python-docx>=1.1.0

# Database
aiosqlite>=0.20.0

# Automation fallback
playwright>=1.48.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# EDI (optional - Phase 2)
# bots-edi>=1.0.0
